split folder from repo
